+++ netharn/__init__.py
@@ -68,7 +68,6 @@ from netharn import criterions
 from netharn import data
 from netharn import device
 from netharn import exceptions
-from netharn import export
 from netharn import fit_harn
 from netharn import hyperparams
 from netharn import initializers
@@ -90,7 +89,7 @@ __all__ = ['Criterion', 'Dynamics', 'FitHarn', 'HiddenFields', 'HiddenShapes',
            'Optimizer', 'OutputShape', 'OutputShapeFor', 'ReceptiveField',
            'ReceptiveFieldFor', 'Scheduler', 'XPU', 'analytic_for', 'api',
            'configure_hacks', 'configure_workdir', 'criterions', 'data',
-           'device', 'exceptions', 'export', 'fit_harn', 'hyperparams',
+           'device', 'exceptions', 'fit_harn', 'hyperparams',
            'initializers', 'layers', 'metrics', 'mixins', 'models', 'monitor',
            'optimizers', 'output_shape_for', 'prefit', 'receptive_field_for',
            'schedulers', 'util']